commit alpha 1.0

+++ Leaps Of Faith/Assets/turretScr.cs
@@ -12,7 +12,7 @@ public class turretScr : MonoBehaviour
     public int xCord;
     public int yCord;
     public int zCord;
-    public int timeMax;
+    public float timeMax;
   
     void Start()
     {
